+++ src/entropy.cpp
@@ -523,15 +523,15 @@ double hastings_correction(int vertex, const Graph &graph, const Blockmodel &blo
     return p_backward / p_forward;
 }
 
-double normalize_mdl_v1(double mdl, int num_edges) {
+double normalize_mdl_v1(double mdl, long num_edges) {
     return mdl / null_mdl_v1(num_edges);
 }
 
-double normalize_mdl_v2(double mdl, int num_vertices, int num_edges) {
+double normalize_mdl_v2(double mdl, long num_vertices, long num_edges) {
     return mdl / null_mdl_v2(num_vertices, num_edges);
 }
 
-double null_mdl_v1(int num_edges) {
+double null_mdl_v1(long num_edges) {
     double log_posterior_p = num_edges * log(1.0 / num_edges);
     double x = 1.0 / num_edges;
     double h = ((1 + x) * log(1 + x)) - (x * log(x));
@@ -539,7 +539,7 @@ double null_mdl_v1(int num_edges) {
     return (num_edges * h) - log_posterior_p;
 }
 
-double null_mdl_v2(int num_vertices, int num_edges) {
+double null_mdl_v2(long num_vertices, long num_edges) {
     double log_posterior_p = num_edges * log(1.0 / num_edges);
     // done calculating log_posterior_probability
     double x = pow(num_vertices, 2) / num_edges;
